removing name as unique key from groups and recipes and refactoring needed services and controllers for that, also adding group information data to recipe

+++ src/main/java/br/ufrgs/foodbook/dao/GroupDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupDao extends JpaRepository<Group, Integer>
+public interface GroupDao extends JpaRepository<Group, Long>
 {
     Group findByName(String groupName);
 }
